removed forceVisible prop, moved floating to its own file
